gh-111968: Use per-thread freelists for generator in free-threading (gh-114189)
