+++ Include/internal/pycore_freelist.h
@@ -19,12 +19,14 @@ extern "C" {
 #  define PyList_MAXFREELIST 80
 #  define PyFloat_MAXFREELIST 100
 #  define PyContext_MAXFREELIST 255
+# define _PyAsyncGen_MAXFREELIST 80
 #else
 #  define PyTuple_NFREELISTS 0
 #  define PyTuple_MAXFREELIST 0
 #  define PyList_MAXFREELIST 0
 #  define PyFloat_MAXFREELIST 0
 #  define PyContext_MAXFREELIST 0
+#  define _PyAsyncGen_MAXFREELIST 0
 #endif
 
 struct _Py_list_state {
@@ -77,12 +79,27 @@ struct _Py_context_state {
 #endif
 };
 
+struct _Py_async_gen_state {
+#ifdef WITH_FREELISTS
+    /* Freelists boost performance 6-10%; they also reduce memory
+       fragmentation, as _PyAsyncGenWrappedValue and PyAsyncGenASend
+       are short-living objects that are instantiated for every
+       __anext__() call. */
+    struct _PyAsyncGenWrappedValue* value_freelist[_PyAsyncGen_MAXFREELIST];
+    int value_numfree;
+
+    struct PyAsyncGenASend* asend_freelist[_PyAsyncGen_MAXFREELIST];
+    int asend_numfree;
+#endif
+};
+
 typedef struct _Py_freelist_state {
     struct _Py_float_state float_state;
     struct _Py_tuple_state tuple_state;
     struct _Py_list_state list_state;
     struct _Py_slice_state slice_state;
     struct _Py_context_state context_state;
+    struct _Py_async_gen_state async_gen_state;
 } _PyFreeListState;
 
 #ifdef __cplusplus
