+++ Python/gc_gil.c
@@ -12,7 +12,6 @@ void
 _PyGC_ClearAllFreeLists(PyInterpreterState *interp)
 {
     _PyDict_ClearFreeList(interp);
-    _PyAsyncGen_ClearFreeLists(interp);
 
     _Py_ClearFreeLists(&interp->freelist_state, 0);
 }
